Add using column number to create_hw_context(

+++ src/runtime_src/core/include/xrt/xrt_hw_context.h
@@ -115,13 +115,13 @@ public:
 
   // Added by xclbin_to_elf
   XRT_API_EXPORT
-  hw_context(const xrt::device& device, uint32_t col_num, access_mode mode);
+  hw_context(const xrt::device& device, uint16_t col_num, access_mode mode);
 
   ///@cond
   // Undocumented construction w/o specifying qos
   // Subject to change in default qos value
   // Added by xclbin_to_elf
-  hw_context(const xrt::device& device, uint32_t col_num)
+  hw_context(const xrt::device& device, uint16_t col_num)
     : hw_context{device, col_num, access_mode::shared}
   {}
   /// @endcond
